+++ package.json
@@ -29,9 +29,9 @@
     "mock": "env-cmd -f .env npm-run-all --parallel start:mock start:react",
     "build": "react-scripts build",
     "lint": "eslint --quiet .",
-    "test": "react-scripts test --env=jest-environment-jsdom-sixteen",
-    "test:ci": "npm test -- --watchAll=false",
-    "test:coverage": "npm test -- --watchAll=false --coverage",
+    "test": "react-scripts test --silent --env=jest-environment-jsdom-sixteen",
+    "test:ci": "npm run test -- --watchAll=false",
+    "test:coverage": "npm run test -- --watchAll=false --coverage",
     "eject": "react-scripts eject",
     "commit:pre": "npm-run-all --parallel test:ci lint"
   },
