+++ src/components/Layout/CenteredContent.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import '@testing-library/jest-dom/extend-expect';
-import { render } from '../../util/TestProvider';
+import { render } from '../../util/TestSetup';
 
 import CenteredContent from './CenteredContent';
